requirements.txt
bottle
pysvn2
whatthepatch
mysql-connector-python
